#83 moved method to another file
Clean up some comments and fixed small bug.
